# docs/swagger.yaml
basePath: /
definitions:
  dto.MovieListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.MovieResponse'
        type: array
      pagination: {}
      success:
        type: boolean
    type: object
  dto.MovieResponse:
    properties:
      casts:
        items:
          type: string
        type: array
      categories:
        items:
          type: string
        type: array
      created_at:
        type: string
      director_name:
        type: string
      duration_in_minute:
        type: integer
      id:
        type: integer
      image:
        type: string
      name:
        type: string
      release_date:
        type: string
      synopsis:
        type: string
      updated_at:
        type: string
    type: object
  dto.MovieSingleResponse:
    properties:
      data:
        $ref: '#/definitions/dto.MovieResponse'
      success:
        type: boolean
    type: object
host: localhost:8081
info:
  contact: {}
  description: API documentation for Tickitz backend application
  license:
    name: MIT
  title: Backend Tickitz API
  version: "1.0"
paths:
  /admin/movies:
    get:
      consumes:
      - application/json
      description: Get paginated list of movies
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Limit per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MovieListResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get list of movies
      tags:
      - Movies
  /admin/movies/{id}:
    get:
      consumes:
      - application/json
      description: Get movie by ID
      parameters:
      - description: Movie ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MovieSingleResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get movie by ID
      tags:
      - Movies
securityDefinitions:
  ApiKeyAuth:
    description: 'Bearer token used for authorization. Example: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"